LunaTranslator/requirements.txt
Deprecated==1.2.13
jaconv==0.3
numpy==1.23.3
onnxruntime==1.12.1
opencv_python==3.4.11.43
pyclipper==1.2.1
PyHook3==1.6.1
PyOpenAL==0.7.11a1
pyperclip==1.8.2
PyQt5==5.15.7
pytest==6.2.4
pywin32==300
QtPy==1.9.0
regex==2021.11.10
requests==2.27.1
selenium==4.5.0
setuptools==56.0.0
Shapely==1.7.1
six==1.15.0
tzlocal==4.1 